third_party/blink/web_tests/external/wpt/mathml/presentation-markup/scripts/subsup-2-expected.txt

This is a testharness.js-based test.
[FAIL] Width of scripted elements
  assert_approx_equals: width is determined by the left/right sides of base/script (+ some space after script) expected 40 +/- 3 but got 45
[FAIL] Height of scripted elements
  assert_approx_equals: msup height is determined by the top/bottom sides of base/scripts expected 40 +/- 5 but got 50
Harness: the test ran to completion.
